requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langchain-groq
